Refactor: move CircularDependencyCheckerType to issue.config.properties
