+++ backend/src/main/java/com/tissue/api/issue/config/properties/CircularDependencyCheckerType.java
@@ -1,4 +1,4 @@
-package com.tissue.api.issue.config;
+package com.tissue.api.issue.config.properties;
 
 public enum CircularDependencyCheckerType {
 	DFS,
